backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langchain==0.3.25
langgraph==0.4.1
langchain-groq==0.3.2
langchain-community==0.3.24
chromadb==0.5.20
pydantic==2.10.0
python-dotenv==1.0.1
sse-starlette==2.1.0
